Algo de tareas hecho

+++ demo/demo/src/main/java/com/appgestion/app/DTO/MaterialDTO.java
@@ -13,7 +13,7 @@ public class MaterialDTO {
 	
 	private String referencia;
 	
-	private int numero_factura;
+	private Long numero_factura;
 
 	private float precio_unitario;
